add data processing, renaming ports, separate backend module (#26)
* add data processing + style fix typescript

* fix styles back

* fix naming ports

* separate backend module
